flags/src/com/android/healthfitness/flags/AconfigFeatureFlagHelper.java
/*
 * Copyright (C) 2024 The Android Open Source Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.android.healthfitness.flags;

/**
 * A helper class to act as the source of truth for whether a feature is enabled or not by taking
 * into account both feature flag and DB flag. See go/hc-aconfig-and-db.
 */
public final class AconfigFeatureFlagHelper {
    /** Returns a boolean indicating whether PHR feature is enabled. */
    public static boolean isPersonalHealthRecordEnabled() {
        return Flags.personalHealthRecord() && Flags.personalHealthRecordDatabase();
    }
}
